Remove unnecessary os.fdopen call (#21)
We don't need to os.fdopen() in order to fstat() an open file.
It's probably a good idea to set the value of fd to None
after closing it to protect against possibility of closing same
fd twice (even though bug where this was happening has already
been fixed). On a busy system the fd number may be reused rather
quickly resulting in inexplicable errors, so let's protect against
this here.