Do not retrieve app's history in debug
This commit adds changes to not retrieve app's history in debug as it's not really useful and can be asked for later explicitly if required which would be quite rare. Secondly app's history has config of user which can contain sensitive information - it cannot be managed in the schema of app's but can be managed here, however seeing that retrieving history has little to no utility, let's remove it from debug.